Ensure path passed to AddEndpoint is not empty
Providing an empty path when registering an endpoint does not cause an
error. However, such endpoint would never work because underneath, it
would end up being "basePath.", and NATS would not allow sending
messages to such subject.

Therefore, as a measure of good coding practice, the function call will
return an error if the path is empty.